+++ src/bencode/BInt.java
@@ -1,16 +1,16 @@
 package bencode;
 
 /**
- * Represents a 32-bit bencoded integer.
+ * Represents a long bencoded integer.
  */
 public class BInt extends BType {
-    private int _value;
-    public BInt( int n ) {
+    private long _value;
+    public BInt( long n ) {
         _value = n;
         _type = Type.INT;
     }
 
-    public Integer value() {
+    public Long value() {
         return _value;
     }
